postid changed to firebase (가 컬렉션에서  자체 제공하는) posts id => delete error resolved

+++ src/components/postDetail/PostDetailBrowse.jsx
@@ -7,26 +7,32 @@ import { styled } from "styled-components";
 
 const PostDetailBrowse = () => {
   console.log("여기는 POSTDETAILBROWSE");
-  const { id } = useParams();
+  const { id } = useParams(); // id === documentId
   const posts = useSelector((state) => state.posts);
-  const post = posts.filter((post) => post.postId === id)[0];
+  const post = posts.filter((post) => post.id === id)[0];
   
   const dispatch = useDispatch();
   const navigate = useNavigate();
+  if(!post) {
+    navigate("/post")
+    return;
+  }
   return (
     <S.PostDetailBox>
-      <p>{post.postId}</p>
+      <p>{post.id}</p>
       <p>{post.postTitle}</p>
       <p>{post.postBody}</p>
       <button
         onClick={async () => {
-          navigate("/post");
+
+          console.log('post.id => ', post.id)
+          
           const postRef = doc(db, "posts", post.id);
           await deleteDoc(postRef);
-          
+          // navigate("/post");
           dispatch({
             type: "DELETE_POST",
-            payload: post.postId,
+            payload: post.id,
           });
         }}
       >
